WIP, will abandon that path
Turns out zo files are really designed to access submodules independently,
so the idea of caching a module's submodules if it's accesses doesn't make
much sense.

Any caching would have to be done at the zo level, which is more involved,
and less likely to be a win, since we'd be increasing memory usage by
potentially a lot, esp. in cases where only small submodules are used, and
not the parent module (e.g. TR's type-decl submodules).

With the current zo loading, the best we can hope for is to warm up OS-level
caches, and find a sweet spot where that effect is larger than the cost of
loading larger zos repeatedly.

So, back to merging clusters of modules and looking for a sweet spot.